Refactoring with BaseSpikeVisual.
